working on the favorite drink display
