+++ models/user.js
@@ -31,7 +31,8 @@ userSchema.set('toObject', {
     let returnJson = {
       _id: ret._id,
       email: ret.email,
-      name: ret.name
+      name: ret.name,
+      drinks: ret.drinks
     }
     return returnJson;
   }
