+++ routes/api.js
@@ -92,6 +92,8 @@ router.get('/drinks/:drinkid', (req, res) => {
 //   })
 // })
 
+
+
 //GET /users/:userid/drinks -- get drinks for one user -- WORKS
 router.get('/drinks', (req, res) => {
   User.findById(req.user._id).populate('drinks').exec((err, user) => {
@@ -99,6 +101,8 @@ router.get('/drinks', (req, res) => {
     res.json(user)
   }) 
 })
+
+
 
 //GET /users/:userid/drinks/:drinkid -- WORKS
 // router.get('/drinks/:drinkid', (req, res) => {
@@ -111,7 +115,7 @@ router.get('/drinks', (req, res) => {
 //POST /users/:userid/drinks -- create new drink -- WORKS, BUT need to figure out how to post ingredients
 router.post('/drinks', (req, res) =>{
   User.findById(req.user._id, function(err, user) {
-    Drink.save(
+    Drink.findById(
       req.body._id, 
       function(err,drink){
           user.drinks.push(drink)
